Undo TCN MSE changes to what is included in the paper. Refactored experiment executable python files into a folder. Grid search py file is moved into TCN folder

+++ tcn/models.py
@@ -63,7 +63,7 @@ class GenericTemporalConvNet(nn.Module):
                     y_ = y_.cuda()
                 optimizer.zero_grad()
                 output = self(x_)
-                loss = loss_function(output.float(), y_.float())
+                loss = loss_function(output, y_)
                 loss.backward()
                 if clip > 0:
                     torch.nn.utils.clip_grad_norm_(self.parameters(), clip)
